:recycle: Refactor IntermediateEvent Events
