+++ src/messages/system_events/index.ts
@@ -6,8 +6,9 @@ export * from './empty_activity_reached';
 export * from './finish_empty_activity';
 export * from './finish_manual_task';
 export * from './finish_user_task';
-export * from './intermediate_event_finished';
-export * from './intermediate_event_triggered';
+export * from './intermediate_catch_event_reached';
+export * from './intermediate_catch_event_finished';
+export * from './intermediate_throw_event_triggered';
 export * from './manual_task_finished';
 export * from './manual_task_reached';
 export * from './process_started';
