+++ src/messages/system_events/intermediate_catch_event_finished.ts
