+++ src/messages/system_events/intermediate_catch_event_reached.ts
@@ -1,0 +1,6 @@
+import {BaseEventMessage} from '../base_event_message';
+
+/**
+ * Represents the message which is send, when an IntermediateCatchEvent has been reached.
+ */
+export class IntermediateCatchEventReachedMessage extends BaseEventMessage {}
