--- src/messages/system_events/intermediate_event_triggered.ts
@@ -1,6 +1,0 @@
-import {BaseEventMessage} from '../base_event_message';
-
-/**
- * Represents the message which is send, when an IntermediateEvent has been triggered.
- */
-export class IntermediateEventTriggeredMessage extends BaseEventMessage {}
