+++ src/messages/system_events/intermediate_throw_event_triggered.ts
@@ -1,0 +1,6 @@
+import {BaseEventMessage} from '../base_event_message';
+
+/**
+ * Represents the message which is send, when an IntermediateThrowEvent has been triggered.
+ */
+export class IntermediateThrowEventTriggeredMessage extends BaseEventMessage {}
